move batch generation logic to generator
